Feat | Extract hooks to a folder
