+++ screens/Home/Home.js
@@ -2,7 +2,7 @@
 import { ScrollView, RefreshControl, Alert, Text } from 'react-native';
 import * as Location from 'expo-location';
 import getTheme from '../../libs/getTheme';
-import useWeather from '../../libs/useWeather';
+import useWeather from '../../hooks/useWeather';
 
 //Context
 import { useState, useEffect, useContext } from 'react';
